% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importElectionWorksheets}
\alias{importElectionWorksheets}
\title{Import election data files.}
\usage{
importElectionWorksheets(filename)
}
\arguments{
\item{filename}{The file to import.}
}
\value{
A dataframe containing the results from each worksheet.
}
\description{
A helper function to convert the Excel files released on Toronto's OpenData site into R dataframes
}
\examples{
\dontrun{
importWorksheets("2010_Mayor.xls")
}
}

