% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toPollGeo}
\alias{toPollGeo}
\title{Geographic data for the polling stations in Toronto municipal elections.}
\format{A data frame with 387422 rows and 12 variables}
\description{
Geocoded poll locations from the 2003, 2006, and
2010 municipal elections in Toronto.
}
\details{
\itemize{
  \item long. The longitude of the polling station
  \item lat. The latitude of the polling station
  \item order. Geocoding of the polling station
  \item hole. Geocoding of the polling station
  \item piece. Geocoding of the polling station
  \item group. Geocoding of the polling station
  \item id. A character string with two characters for ward and three for area
  \item year. The year in which the polling station was used
  \item ward. The ward of the polling station
  \item area. The area of the polling station
  \item ward name. The name of the ward containing the polling station
  ...
}
}

